requirements.txt
dlib
pillow
scikit-image
matplotlib
numpy
imageio
matplotlib
mlxtend
opencv-python
